Make the maximum RTP packet size for the H.264 packetizer configurable
This can help with mitigating packet drops in networks with small
MTU. At the same time, we're also changing the default limit
for H.264 RTP packets to 1200 bytes. This value is commonly used
by other RTP implementations as a reasonable trade-off to fit
the generated packets into MTU limits of most networks across the
Internet.